requirements.txt
requests
python-dotenv
django
pandas
pyopenstates
